backend/src/convo-summary/requirements.txt
fastapi==0.115.12
pydantic==2.11.1
Requests==2.32.3
python-multipart==0.0.20
uvicorn==0.34.0
python-dotenv==1.0.0
mysqlclient
PyYAML